+++ prefix_benchmark_test.go
@@ -91,7 +91,7 @@ func BenchmarkAcquireIPCockroach(b *testing.B) {
 	benchmarkAcquireIP(ipam, "10.0.0.0/16", b)
 }
 
-func benchmarkAcquireChildPrefix(parentLength, childLength int, b *testing.B) {
+func benchmarkAcquireChildPrefix(parentLength, childLength uint8, b *testing.B) {
 	ipam := New()
 	p, err := ipam.NewPrefix(fmt.Sprintf("192.168.0.0/%d", parentLength))
 	if err != nil {
